soft delete topic

+++ app/Models/Classroom.php
@@ -5,11 +5,12 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Classroom extends Model
 {
 
-    use HasFactory;
+    use HasFactory , SoftDeletes;
 
     //to  enhance code from maintanace
     public static string $disk='public';
@@ -37,7 +38,9 @@ public static function  uploadCoverImage($file){
 }
 
 public static function  deleteCoverImage($path){
+if($path || Storage::disk(Classroom::$disk)->exists($path)){
     return Storage::disk(Classroom::$disk)->delete($path);
+}
 
 }
 }
